Revert "Navigation enhancements: Bottom Navigation, screen transitions, back button (#670)"
This reverts commit 36bde4534bc7ee61ef63973847d66ab8804495c1.
