+++ app/src/main/java/com/jerboa/MainActivity.kt
@@ -1,18 +1,27 @@
 package com.jerboa
 
 import android.app.Application
+import android.content.Intent
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
+import android.net.Uri
+import android.os.Build
 import android.os.Bundle
+import android.util.Patterns
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.platform.LocalContext
-import com.google.accompanist.navigation.animation.AnimatedNavHost
-import com.google.accompanist.navigation.animation.rememberAnimatedNavController
+import androidx.navigation.NavType
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import androidx.navigation.navArgument
+import androidx.navigation.navDeepLink
+import arrow.core.Either
 import com.jerboa.db.AccountRepository
 import com.jerboa.db.AccountViewModel
 import com.jerboa.db.AccountViewModelFactory
@@ -20,52 +29,42 @@ import com.jerboa.db.AppDB
 import com.jerboa.db.AppSettingsRepository
 import com.jerboa.db.AppSettingsViewModel
 import com.jerboa.db.AppSettingsViewModelFactory
-import com.jerboa.nav.aboutScreen
-import com.jerboa.nav.accountSettingsScreen
-import com.jerboa.nav.commentEditScreen
-import com.jerboa.nav.commentReplyScreen
-import com.jerboa.nav.commentReportScreen
-import com.jerboa.nav.commentScreen
-import com.jerboa.nav.communityListScreen
-import com.jerboa.nav.communityScreen
-import com.jerboa.nav.communityScreenFromUrl
-import com.jerboa.nav.communitySideBarScreen
-import com.jerboa.nav.createPostScreen
-import com.jerboa.nav.defaultEnterTransition
-import com.jerboa.nav.defaultExitTransition
-import com.jerboa.nav.defaultPopEnterTransition
-import com.jerboa.nav.defaultPopExitTransition
-import com.jerboa.nav.homeRoutePattern
-import com.jerboa.nav.homeScreen
-import com.jerboa.nav.inboxScreen
-import com.jerboa.nav.loginScreen
-import com.jerboa.nav.lookAndFeelScreen
-import com.jerboa.nav.postEditScreen
-import com.jerboa.nav.postReportScreen
-import com.jerboa.nav.postScreen
-import com.jerboa.nav.privateMessageReplyScreen
-import com.jerboa.nav.profileScreen
-import com.jerboa.nav.profileScreenFromUrl
-import com.jerboa.nav.settingsScreen
-import com.jerboa.nav.siteSideBarScreen
+import com.jerboa.ui.components.comment.edit.CommentEditActivity
 import com.jerboa.ui.components.comment.edit.CommentEditViewModel
+import com.jerboa.ui.components.comment.reply.CommentReplyActivity
 import com.jerboa.ui.components.comment.reply.CommentReplyViewModel
 import com.jerboa.ui.components.common.MarkdownHelper
 import com.jerboa.ui.components.common.ShowChangelog
 import com.jerboa.ui.components.common.getCurrentAccount
 import com.jerboa.ui.components.common.getCurrentAccountSync
+import com.jerboa.ui.components.community.CommunityActivity
 import com.jerboa.ui.components.community.CommunityViewModel
+import com.jerboa.ui.components.community.list.CommunityListActivity
 import com.jerboa.ui.components.community.list.CommunityListViewModel
+import com.jerboa.ui.components.community.sidebar.CommunitySidebarActivity
 import com.jerboa.ui.components.home.*
+import com.jerboa.ui.components.inbox.InboxActivity
 import com.jerboa.ui.components.inbox.InboxViewModel
+import com.jerboa.ui.components.login.LoginActivity
 import com.jerboa.ui.components.login.LoginViewModel
+import com.jerboa.ui.components.person.PersonProfileActivity
 import com.jerboa.ui.components.person.PersonProfileViewModel
+import com.jerboa.ui.components.post.PostActivity
 import com.jerboa.ui.components.post.PostViewModel
+import com.jerboa.ui.components.post.create.CreatePostActivity
 import com.jerboa.ui.components.post.create.CreatePostViewModel
+import com.jerboa.ui.components.post.edit.PostEditActivity
 import com.jerboa.ui.components.post.edit.PostEditViewModel
+import com.jerboa.ui.components.privatemessage.PrivateMessageReplyActivity
 import com.jerboa.ui.components.report.CreateReportViewModel
+import com.jerboa.ui.components.report.comment.CreateCommentReportActivity
+import com.jerboa.ui.components.report.post.CreatePostReportActivity
+import com.jerboa.ui.components.settings.SettingsActivity
+import com.jerboa.ui.components.settings.about.AboutActivity
+import com.jerboa.ui.components.settings.account.AccountSettingsActivity
 import com.jerboa.ui.components.settings.account.AccountSettingsViewModel
 import com.jerboa.ui.components.settings.account.AccountSettingsViewModelFactory
+import com.jerboa.ui.components.settings.lookandfeel.LookAndFeelActivity
 import com.jerboa.ui.theme.JerboaTheme
 
 class JerboaApplication : Application() {
@@ -89,7 +88,7 @@ class MainActivity : ComponentActivity() {
     private val commentEditViewModel by viewModels<CommentEditViewModel>()
     private val postEditViewModel by viewModels<PostEditViewModel>()
     private val createReportViewModel by viewModels<CreateReportViewModel>()
-    private val accountSettingsViewModel by viewModels<AccountSettingsViewModel> {
+    private val accountSettingsViewModel by viewModels<AccountSettingsViewModel>() {
         AccountSettingsViewModelFactory((application as JerboaApplication).accountRepository)
     }
     private val accountViewModel: AccountViewModel by viewModels {
@@ -99,7 +98,6 @@ class MainActivity : ComponentActivity() {
         AppSettingsViewModelFactory((application as JerboaApplication).appSettingsRepository)
     }
 
-    @OptIn(ExperimentalAnimationApi::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -115,7 +113,7 @@ class MainActivity : ComponentActivity() {
             JerboaTheme(
                 appSettings = appSettings,
             ) {
-                val navController = rememberAnimatedNavController()
+                val navController = rememberNavController()
                 val ctx = LocalContext.current
 
                 MarkdownHelper.init(
@@ -126,221 +124,522 @@ class MainActivity : ComponentActivity() {
 
                 ShowChangelog(appSettingsViewModel = appSettingsViewModel)
 
-                AnimatedNavHost(
+                NavHost(
                     navController = navController,
-                    startDestination = homeRoutePattern,
-                    enterTransition = { defaultEnterTransition },
-                    exitTransition = { defaultExitTransition },
-                    popEnterTransition = { defaultPopEnterTransition },
-                    popExitTransition = { defaultPopExitTransition },
+                    startDestination = "home",
                 ) {
-                    loginScreen(
-                        navController = navController,
-                        loginViewModel = loginViewModel,
-                        accountViewModel = accountViewModel,
-                        siteViewModel = siteViewModel,
-                        homeViewModel = homeViewModel,
-                    )
+                    composable(
+                        route = "login",
+                        deepLinks = DEFAULT_LEMMY_INSTANCES.map { instance ->
+                            navDeepLink { uriPattern = "$instance/login" }
+                        },
+                    ) {
+                        LoginActivity(
+                            navController = navController,
+                            loginViewModel = loginViewModel,
+                            accountViewModel = accountViewModel,
+                            siteViewModel = siteViewModel,
+                            homeViewModel = homeViewModel,
+                        )
+                    }
+                    composable(
+                        route = "home",
+                    ) {
+                        HomeActivity(
+                            navController = navController,
+                            homeViewModel = homeViewModel,
+                            accountViewModel = accountViewModel,
+                            siteViewModel = siteViewModel,
+                            postEditViewModel = postEditViewModel,
+                            appSettingsViewModel = appSettingsViewModel,
+                            showVotingArrowsInListView = appSettings?.showVotingArrowsInListView ?: true,
+                        )
+                    }
+                    composable(
+                        route = "community/{id}",
+                        arguments = listOf(
+                            navArgument("id") {
+                                type = NavType.IntType
+                            },
+                        ),
+                    ) {
+                        LaunchedEffect(Unit) {
+                            val communityId = it.arguments?.getInt("id")!!
+                            val idOrName = Either.Left(communityId)
 
-                    homeScreen(
-                        navController = navController,
-                        homeViewModel = homeViewModel,
-                        accountViewModel = accountViewModel,
-                        siteViewModel = siteViewModel,
-                        postEditViewModel = postEditViewModel,
-                        appSettingsViewModel = appSettingsViewModel,
-                        communityListViewModel = communityListViewModel,
-                        inboxViewModel = inboxViewModel,
-                        commentReplyViewModel = commentReplyViewModel,
-                        personProfileViewModel = personProfileViewModel,
-                        commentEditViewModel = commentEditViewModel,
-                        appSettings = appSettings,
-                    )
+                            communityViewModel.fetchCommunity(
+                                idOrName = idOrName,
+                                auth = account?.jwt,
+                            )
 
-                    communityScreen(
-                        navController = navController,
-                        communityViewModel = communityViewModel,
-                        accountViewModel = accountViewModel,
-                        postEditViewModel = postEditViewModel,
-                        communityListViewModel = communityListViewModel,
-                        appSettingsViewModel = appSettingsViewModel,
-                        siteViewModel = siteViewModel,
-                        account = account,
-                        appSettings = appSettings,
-                        ctx = ctx,
-                    )
+                            communityViewModel.fetchPosts(
+                                communityIdOrName = idOrName,
+                                account = account,
+                                clear = true,
+                                ctx = ctx,
+                            )
+                        }
 
-                    communityScreenFromUrl(
-                        navController = navController,
-                        communityViewModel = communityViewModel,
-                        accountViewModel = accountViewModel,
-                        postEditViewModel = postEditViewModel,
-                        communityListViewModel = communityListViewModel,
-                        appSettingsViewModel = appSettingsViewModel,
-                        siteViewModel = siteViewModel,
-                        account = account,
-                        appSettings = appSettings,
-                        ctx = ctx,
-                    )
+                        CommunityActivity(
+                            navController = navController,
+                            communityViewModel = communityViewModel,
+                            accountViewModel = accountViewModel,
+                            homeViewModel = homeViewModel,
+                            postEditViewModel = postEditViewModel,
+                            communityListViewModel = communityListViewModel,
+                            appSettingsViewModel = appSettingsViewModel,
+                            showVotingArrowsInListView = appSettings?.showVotingArrowsInListView ?: true,
+                            siteViewModel = siteViewModel,
+                        )
+                    }
+                    // Only necessary for community deeplinks
+                    composable(
+                        route = "{instance}/c/{name}",
+                        deepLinks = listOf(
+                            navDeepLink { uriPattern = "{instance}/c/{name}" },
+                        ),
+                        arguments = listOf(
+                            navArgument("name") {
+                                type = NavType.StringType
+                            },
+                            navArgument("instance") {
+                                type = NavType.StringType
+                            },
+                        ),
+                    ) {
+                        LaunchedEffect(Unit) {
+                            val name = it.arguments?.getString("name")!!
+                            val instance = it.arguments?.getString("instance")!!
+                            val idOrName = Either.Right("$name@$instance")
 
-                    profileScreen(
-                        navController = navController,
-                        personProfileViewModel = personProfileViewModel,
-                        accountViewModel = accountViewModel,
-                        commentEditViewModel = commentEditViewModel,
-                        commentReplyViewModel = commentReplyViewModel,
-                        postEditViewModel = postEditViewModel,
-                        appSettingsViewModel = appSettingsViewModel,
-                        siteViewModel = siteViewModel,
-                        account = account,
-                        appSettings = appSettings,
-                    )
+                            communityViewModel.fetchCommunity(
+                                idOrName = idOrName,
+                                auth = account?.jwt,
+                            )
 
-                    profileScreenFromUrl(
-                        navController = navController,
-                        personProfileViewModel = personProfileViewModel,
-                        accountViewModel = accountViewModel,
-                        commentEditViewModel = commentEditViewModel,
-                        commentReplyViewModel = commentReplyViewModel,
-                        postEditViewModel = postEditViewModel,
-                        appSettingsViewModel = appSettingsViewModel,
-                        siteViewModel = siteViewModel,
-                        account = account,
-                        appSettings = appSettings,
-                        ctx = ctx,
-                    )
+                            communityViewModel.fetchPosts(
+                                communityIdOrName = idOrName,
+                                account = account,
+                                clear = true,
+                                ctx = ctx,
+                            )
+                        }
 
-                    communityListScreen(
-                        navController = navController,
-                        accountViewModel = accountViewModel,
-                        communityListViewModel = communityListViewModel,
-                        siteViewModel = siteViewModel,
-                    )
+                        CommunityActivity(
+                            navController = navController,
+                            communityViewModel = communityViewModel,
+                            communityListViewModel = communityListViewModel,
+                            accountViewModel = accountViewModel,
+                            homeViewModel = homeViewModel,
+                            postEditViewModel = postEditViewModel,
+                            appSettingsViewModel = appSettingsViewModel,
+                            showVotingArrowsInListView = appSettings?.showVotingArrowsInListView ?: true,
+                            siteViewModel = siteViewModel,
+                        )
+                    }
+                    composable(
+                        route = "profile/{id}?saved={saved}",
+                        arguments = listOf(
+                            navArgument("id") {
+                                type = NavType.IntType
+                            },
+                            navArgument("saved") {
+                                defaultValue = false
+                                type = NavType.BoolType
+                            },
+                        ),
+                    ) {
+                        val savedMode = it.arguments?.getBoolean("saved")!!
 
-                    createPostScreen(
-                        navController = navController,
-                        accountViewModel = accountViewModel,
-                        createPostViewModel = createPostViewModel,
-                        communityListViewModel = communityListViewModel,
-                        ctx = ctx,
-                    )
+                        LaunchedEffect(Unit) {
+                            val personId = it.arguments?.getInt("id")!!
+                            val idOrName = Either.Left(personId)
 
-                    inboxScreen(
-                        navController = navController,
-                        inboxViewModel = inboxViewModel,
-                        accountViewModel = accountViewModel,
-                        homeViewModel = homeViewModel,
-                        commentReplyViewModel = commentReplyViewModel,
-                        siteViewModel = siteViewModel,
-                        account = account,
-                        ctx = ctx,
-                    )
+                            personProfileViewModel.fetchPersonDetails(
+                                idOrName = idOrName,
+                                account = account,
+                                clearPersonDetails = true,
+                                clearPostsAndComments = true,
+                                ctx = ctx,
+                                changeSavedOnly = savedMode,
+                            )
+                        }
 
-                    postScreen(
-                        navController = navController,
-                        postViewModel = postViewModel,
-                        accountViewModel = accountViewModel,
-                        commentEditViewModel = commentEditViewModel,
-                        commentReplyViewModel = commentReplyViewModel,
-                        postEditViewModel = postEditViewModel,
-                        appSettingsViewModel = appSettingsViewModel,
-                        siteViewModel = siteViewModel,
-                        account = account,
-                        appSettings = appSettings,
-                        ctx = ctx,
-                    )
+                        PersonProfileActivity(
+                            savedMode = savedMode,
+                            navController = navController,
+                            personProfileViewModel = personProfileViewModel,
+                            accountViewModel = accountViewModel,
+                            homeViewModel = homeViewModel,
+                            commentEditViewModel = commentEditViewModel,
+                            commentReplyViewModel = commentReplyViewModel,
+                            postEditViewModel = postEditViewModel,
+                            appSettingsViewModel = appSettingsViewModel,
+                            showVotingArrowsInListView = appSettings?.showVotingArrowsInListView ?: true,
+                            siteViewModel = siteViewModel,
+                        )
+                    }
+                    // Necessary for deep links
+                    composable(
+                        route = "{instance}/u/{name}",
+                        deepLinks = listOf(
+                            navDeepLink { uriPattern = "{instance}/u/{name}" },
+                        ),
+                        arguments = listOf(
+                            navArgument("name") {
+                                type = NavType.StringType
+                            },
+                            navArgument("instance") {
+                                type = NavType.StringType
+                            },
+                        ),
+                    ) {
+                        LaunchedEffect(Unit) {
+                            val name = it.arguments?.getString("name")!!
+                            val instance = it.arguments?.getString("instance")!!
+                            val idOrName = Either.Right("$name@$instance")
 
-                    commentScreen(
-                        navController = navController,
-                        postViewModel = postViewModel,
-                        accountViewModel = accountViewModel,
-                        commentEditViewModel = commentEditViewModel,
-                        commentReplyViewModel = commentReplyViewModel,
-                        postEditViewModel = postEditViewModel,
-                        appSettingsViewModel = appSettingsViewModel,
-                        siteViewModel = siteViewModel,
-                        account = account,
-                        appSettings = appSettings,
-                        ctx = ctx,
-                    )
+                            personProfileViewModel.fetchPersonDetails(
+                                idOrName = idOrName,
+                                account = account,
+                                clearPersonDetails = true,
+                                clearPostsAndComments = true,
+                                ctx = ctx,
+                            )
+                        }
 
-                    commentReplyScreen(
-                        commentReplyViewModel = commentReplyViewModel,
-                        postViewModel = postViewModel,
-                        accountViewModel = accountViewModel,
-                        personProfileViewModel = personProfileViewModel,
-                        navController = navController,
-                        siteViewModel = siteViewModel,
-                    )
+                        PersonProfileActivity(
+                            savedMode = false,
+                            navController = navController,
+                            personProfileViewModel = personProfileViewModel,
+                            accountViewModel = accountViewModel,
+                            homeViewModel = homeViewModel,
+                            commentEditViewModel = commentEditViewModel,
+                            commentReplyViewModel = commentReplyViewModel,
+                            postEditViewModel = postEditViewModel,
+                            appSettingsViewModel = appSettingsViewModel,
+                            showVotingArrowsInListView = appSettings?.showVotingArrowsInListView ?: true,
+                            siteViewModel = siteViewModel,
+                        )
+                    }
+                    composable(
+                        route = "communityList?select={select}",
+                        arguments = listOf(
+                            navArgument("select") {
+                                defaultValue = false
+                                type = NavType.BoolType
+                            },
+                        ),
+                    ) {
+                        // Whenever navigating here, reset the list with your followed communities
+                        communityListViewModel.setCommunityListFromFollowed(siteViewModel)
 
-                    siteSideBarScreen(
-                        siteViewModel = siteViewModel,
-                        navController = navController,
-                    )
+                        CommunityListActivity(
+                            navController = navController,
+                            accountViewModel = accountViewModel,
+                            homeViewModel = homeViewModel,
+                            appSettingsViewModel = appSettingsViewModel,
+                            communityListViewModel = communityListViewModel,
+                            selectMode = it.arguments?.getBoolean("select")!!,
+                        )
+                    }
+                    composable(
+                        route = "createPost",
+                        deepLinks = listOf(
+                            navDeepLink { mimeType = "text/plain" },
+                            navDeepLink { mimeType = "image/*" },
+                        ),
+                    ) {
+                        val activity = ctx.findActivity()
+                        val text = activity?.intent?.getStringExtra(Intent.EXTRA_TEXT) ?: ""
+                        val image =
+                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+                                activity?.intent?.getParcelableExtra(
+                                    Intent.EXTRA_STREAM,
+                                    Uri::class.java,
+                                )
+                            } else {
+                                @Suppress("DEPRECATION")
+                                activity?.intent?.getParcelableExtra(Intent.EXTRA_STREAM) as? Uri
+                            }
+                        // url and body will be empty everytime except when there is EXTRA TEXT in the intent
+                        var url = ""
+                        var body = ""
+                        if (Patterns.WEB_URL.matcher(text).matches()) {
+                            url = text
+                        } else {
+                            body = text
+                        }
 
-                    communitySideBarScreen(
-                        communityViewModel = communityViewModel,
-                        navController = navController,
-                    )
+                        CreatePostActivity(
+                            navController = navController,
+                            accountViewModel = accountViewModel,
+                            createPostViewModel = createPostViewModel,
+                            communityListViewModel = communityListViewModel,
+                            _url = url,
+                            _body = body,
+                            _image = image,
+                        )
+                        activity?.intent?.replaceExtras(Bundle())
+                    }
+                    composable(
+                        route = "inbox",
+                        deepLinks = DEFAULT_LEMMY_INSTANCES.map { instance ->
+                            navDeepLink { uriPattern = "$instance/inbox" }
+                        },
+                    ) {
+                        if (account != null) {
+                            LaunchedEffect(Unit) {
+                                inboxViewModel.fetchReplies(
+                                    account = account,
+                                    clear = true,
+                                    ctx = ctx,
+                                )
+                                inboxViewModel.fetchPersonMentions(
+                                    account = account,
+                                    clear = true,
+                                    ctx = ctx,
+                                )
+                                inboxViewModel.fetchPrivateMessages(
+                                    account = account,
+                                    clear = true,
+                                    ctx = ctx,
+                                )
+                            }
+                        }
 
-                    commentEditScreen(
-                        commentEditViewModel = commentEditViewModel,
-                        accountViewModel = accountViewModel,
-                        navController = navController,
-                        personProfileViewModel = personProfileViewModel,
-                        postViewModel = postViewModel,
-                    )
-
-                    postEditScreen(
-                        postEditViewModel = postEditViewModel,
-                        communityViewModel = communityViewModel,
-                        accountViewModel = accountViewModel,
-                        navController = navController,
-                        personProfileViewModel = personProfileViewModel,
-                        postViewModel = postViewModel,
-                        homeViewModel = homeViewModel,
-                    )
-
-                    privateMessageReplyScreen(
-                        inboxViewModel = inboxViewModel,
-                        accountViewModel = accountViewModel,
-                        navController = navController,
-                        siteViewModel = siteViewModel,
-                    )
-
-                    commentReportScreen(
-                        createReportViewModel = createReportViewModel,
-                        accountViewModel = accountViewModel,
-                        navController = navController,
-                    )
-
-                    postReportScreen(
-                        createReportViewModel = createReportViewModel,
-                        accountViewModel = accountViewModel,
-                        navController = navController,
-                    )
-
-                    settingsScreen(
-                        navController = navController,
-                        accountViewModel = accountViewModel,
-                    )
-
-                    lookAndFeelScreen(
-                        navController = navController,
-                        appSettingsViewModel = appSettingsViewModel,
-                    )
-
-                    accountSettingsScreen(
-                        navController = navController,
-                        accountViewModel = accountViewModel,
-                        siteViewModel = siteViewModel,
-                        accountSettingsViewModel = accountSettingsViewModel,
-                    )
-
-                    aboutScreen(
-                        navController = navController,
-                        appSettings = appSettings,
-                    )
+                        InboxActivity(
+                            navController = navController,
+                            appSettingsViewModel = appSettingsViewModel,
+                            inboxViewModel = inboxViewModel,
+                            accountViewModel = accountViewModel,
+                            homeViewModel = homeViewModel,
+                            commentReplyViewModel = commentReplyViewModel,
+                            siteViewModel = siteViewModel,
+                        )
+                    }
+                    composable(
+                        route = "post/{id}",
+                        deepLinks = DEFAULT_LEMMY_INSTANCES.map { instance ->
+                            navDeepLink { uriPattern = "$instance/post/{id}" }
+                        },
+                        arguments = listOf(
+                            navArgument("id") {
+                                type = NavType.IntType
+                            },
+                        ),
+                    ) {
+                        LaunchedEffect(Unit) {
+                            val postId = it.arguments?.getInt("id")!!
+                            postViewModel.fetchPost(
+                                id = Either.Left(postId),
+                                account = account,
+                                clearPost = true,
+                                clearComments = true,
+                                ctx = ctx,
+                            )
+                        }
+                        PostActivity(
+                            postViewModel = postViewModel,
+                            accountViewModel = accountViewModel,
+                            commentEditViewModel = commentEditViewModel,
+                            commentReplyViewModel = commentReplyViewModel,
+                            postEditViewModel = postEditViewModel,
+                            navController = navController,
+                            showCollapsedCommentContent = appSettings?.showCollapsedCommentContent ?: false,
+                            showActionBarByDefault = appSettings?.showCommentActionBarByDefault ?: true,
+                            showVotingArrowsInListView = appSettings?.showVotingArrowsInListView ?: true,
+                            onClickSortType = { commentSortType ->
+                                val postId = it.arguments?.getInt("id")!!
+                                postViewModel.fetchPost(
+                                    id = Either.Left(postId),
+                                    account = account,
+                                    clearPost = false,
+                                    clearComments = true,
+                                    ctx = ctx,
+                                    changeSortType = commentSortType,
+                                )
+                            },
+                            selectedSortType = postViewModel.sortType.value,
+                            siteViewModel = siteViewModel,
+                        )
+                    }
+                    composable(
+                        route = "comment/{id}",
+                        deepLinks = DEFAULT_LEMMY_INSTANCES.map { instance ->
+                            navDeepLink { uriPattern = "$instance/comment/{id}" }
+                        },
+                        arguments = listOf(
+                            navArgument("id") {
+                                type = NavType.IntType
+                            },
+                        ),
+                    ) {
+                        LaunchedEffect(Unit) {
+                            val commentId = it.arguments?.getInt("id")!!
+                            postViewModel.fetchPost(
+                                id = Either.Right(commentId),
+                                account = account,
+                                clearPost = true,
+                                clearComments = true,
+                                ctx = ctx,
+                            )
+                        }
+                        PostActivity(
+                            postViewModel = postViewModel,
+                            accountViewModel = accountViewModel,
+                            commentEditViewModel = commentEditViewModel,
+                            commentReplyViewModel = commentReplyViewModel,
+                            postEditViewModel = postEditViewModel,
+                            navController = navController,
+                            showCollapsedCommentContent = appSettings?.showCollapsedCommentContent ?: false,
+                            showActionBarByDefault = appSettings?.showCommentActionBarByDefault ?: true,
+                            showVotingArrowsInListView = appSettings?.showVotingArrowsInListView ?: true,
+                            onClickSortType = { commentSortType ->
+                                val commentId = it.arguments?.getInt("id")!!
+                                postViewModel.fetchPost(
+                                    id = Either.Right(commentId),
+                                    account = account,
+                                    clearPost = false,
+                                    clearComments = true,
+                                    ctx = ctx,
+                                    changeSortType = commentSortType,
+                                )
+                            },
+                            selectedSortType = postViewModel.sortType.value,
+                            siteViewModel = siteViewModel,
+                        )
+                    }
+                    composable(
+                        route = "commentReply",
+                    ) {
+                        CommentReplyActivity(
+                            commentReplyViewModel = commentReplyViewModel,
+                            postViewModel = postViewModel,
+                            accountViewModel = accountViewModel,
+                            personProfileViewModel = personProfileViewModel,
+                            navController = navController,
+                            siteViewModel = siteViewModel,
+                        )
+                    }
+                    composable(
+                        route = "siteSidebar",
+                    ) {
+                        SiteSidebarActivity(
+                            siteViewModel = siteViewModel,
+                            navController = navController,
+                        )
+                    }
+                    composable(
+                        route = "communitySidebar",
+                    ) {
+                        CommunitySidebarActivity(
+                            communityViewModel = communityViewModel,
+                            navController = navController,
+                        )
+                    }
+                    composable(
+                        route = "commentEdit",
+                    ) {
+                        CommentEditActivity(
+                            commentEditViewModel = commentEditViewModel,
+                            accountViewModel = accountViewModel,
+                            navController = navController,
+                            personProfileViewModel = personProfileViewModel,
+                            postViewModel = postViewModel,
+                        )
+                    }
+                    composable(
+                        route = "postEdit",
+                    ) {
+                        PostEditActivity(
+                            postEditViewModel = postEditViewModel,
+                            communityViewModel = communityViewModel,
+                            accountViewModel = accountViewModel,
+                            navController = navController,
+                            personProfileViewModel = personProfileViewModel,
+                            postViewModel = postViewModel,
+                            homeViewModel = homeViewModel,
+                        )
+                    }
+                    composable(
+                        route = "privateMessageReply",
+                    ) {
+                        PrivateMessageReplyActivity(
+                            inboxViewModel = inboxViewModel,
+                            accountViewModel = accountViewModel,
+                            navController = navController,
+                            siteViewModel = siteViewModel,
+                        )
+                    }
+                    composable(
+                        route = "commentReport/{id}",
+                        arguments = listOf(
+                            navArgument("id") {
+                                type = NavType.IntType
+                            },
+                        ),
+                    ) {
+                        createReportViewModel.setCommentId(it.arguments?.getInt("id")!!)
+                        CreateCommentReportActivity(
+                            createReportViewModel = createReportViewModel,
+                            accountViewModel = accountViewModel,
+                            navController = navController,
+                        )
+                    }
+                    composable(
+                        route = "postReport/{id}",
+                        arguments = listOf(
+                            navArgument("id") {
+                                type = NavType.IntType
+                            },
+                        ),
+                    ) {
+                        createReportViewModel.setPostId(it.arguments?.getInt("id")!!)
+                        CreatePostReportActivity(
+                            createReportViewModel = createReportViewModel,
+                            accountViewModel = accountViewModel,
+                            navController = navController,
+                        )
+                    }
+                    composable(
+                        route = "settings",
+                    ) {
+                        SettingsActivity(
+                            navController = navController,
+                            accountViewModel = accountViewModel,
+                        )
+                    }
+                    composable(
+                        route = "lookAndFeel",
+                    ) {
+                        LookAndFeelActivity(
+                            navController = navController,
+                            appSettingsViewModel = appSettingsViewModel,
+                        )
+                    }
+                    composable(
+                        route = "accountSettings",
+                        deepLinks = DEFAULT_LEMMY_INSTANCES.map { instance ->
+                            navDeepLink { uriPattern = "$instance/settings" }
+                        },
+                    ) {
+                        AccountSettingsActivity(
+                            navController = navController,
+                            accountViewModel = accountViewModel,
+                            siteViewModel = siteViewModel,
+                            accountSettingsViewModel = accountSettingsViewModel,
+                        )
+                    }
+                    composable(
+                        route = "about",
+                    ) {
+                        AboutActivity(
+                            navController = navController,
+                            useCustomTabs = appSettings?.useCustomTabs ?: true,
+                            usePrivateTabs = appSettings?.usePrivateTabs ?: false,
+                        )
+                    }
                 }
             }
         }
