+++ app/src/main/java/com/jerboa/ui/components/community/CommunityActivity.kt
@@ -33,9 +33,11 @@ import com.jerboa.db.AccountViewModel
 import com.jerboa.db.AppSettingsViewModel
 import com.jerboa.loginFirstToast
 import com.jerboa.scrollToTop
+import com.jerboa.ui.components.common.BottomAppBarAll
 import com.jerboa.ui.components.common.getCurrentAccount
 import com.jerboa.ui.components.common.getPostViewMode
 import com.jerboa.ui.components.community.list.CommunityListViewModel
+import com.jerboa.ui.components.home.HomeViewModel
 import com.jerboa.ui.components.home.SiteViewModel
 import com.jerboa.ui.components.post.PostListings
 import com.jerboa.ui.components.post.edit.PostEditViewModel
@@ -47,6 +49,7 @@ fun CommunityActivity(
     communityViewModel: CommunityViewModel,
     communityListViewModel: CommunityListViewModel,
     accountViewModel: AccountViewModel,
+    homeViewModel: HomeViewModel,
     postEditViewModel: PostEditViewModel,
     appSettingsViewModel: AppSettingsViewModel,
     showVotingArrowsInListView: Boolean,
@@ -247,6 +250,35 @@ fun CommunityActivity(
                     contentDescription = stringResource(R.string.floating_createPost),
                 )
             }
+        },
+        bottomBar = {
+            BottomAppBarAll(
+                showBottomNav = appSettingsViewModel.appSettings.value?.showBottomNav,
+                screen = "communityList",
+                unreadCounts = homeViewModel.unreadCountResponse,
+                onClickProfile = {
+                    account?.id?.also {
+                        navController.navigate(route = "profile/$it")
+                    } ?: run {
+                        loginFirstToast(ctx)
+                    }
+                },
+                onClickInbox = {
+                    account?.also {
+                        navController.navigate(route = "inbox")
+                    } ?: run {
+                        loginFirstToast(ctx)
+                    }
+                },
+                onClickSaved = {
+                    account?.id?.also {
+                        navController.navigate(route = "profile/$it?saved=${true}")
+                    } ?: run {
+                        loginFirstToast(ctx)
+                    }
+                },
+                navController = navController,
+            )
         },
     )
 }
