+++ app/src/main/java/com/jerboa/ui/components/community/list/CommunityListActivity.kt
@@ -22,7 +22,11 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavController
 import com.jerboa.DEBOUNCE_DELAY
 import com.jerboa.db.AccountViewModel
+import com.jerboa.db.AppSettingsViewModel
+import com.jerboa.loginFirstToast
+import com.jerboa.ui.components.common.BottomAppBarAll
 import com.jerboa.ui.components.common.getCurrentAccount
+import com.jerboa.ui.components.home.HomeViewModel
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
@@ -34,6 +38,8 @@ fun CommunityListActivity(
     navController: NavController,
     communityListViewModel: CommunityListViewModel,
     accountViewModel: AccountViewModel,
+    homeViewModel: HomeViewModel,
+    appSettingsViewModel: AppSettingsViewModel,
     selectMode: Boolean = false,
 ) {
     Log.d("jerboa", "got to community list activity")
@@ -84,6 +90,35 @@ fun CommunityListActivity(
                             .imePadding(),
                     )
                 }
+            },
+            bottomBar = {
+                BottomAppBarAll(
+                    showBottomNav = appSettingsViewModel.appSettings.value?.showBottomNav,
+                    screen = "communityList",
+                    unreadCounts = homeViewModel.unreadCountResponse,
+                    onClickProfile = {
+                        account?.id?.also {
+                            navController.navigate(route = "profile/$it")
+                        } ?: run {
+                            loginFirstToast(ctx)
+                        }
+                    },
+                    onClickInbox = {
+                        account?.also {
+                            navController.navigate(route = "inbox")
+                        } ?: run {
+                            loginFirstToast(ctx)
+                        }
+                    },
+                    onClickSaved = {
+                        account?.id?.also {
+                            navController.navigate(route = "profile/$it?saved=${true}")
+                        } ?: run {
+                            loginFirstToast(ctx)
+                        }
+                    },
+                    navController = navController,
+                )
             },
         )
     }
