+++ app/src/main/java/com/jerboa/ui/components/home/HomeActivity.kt
@@ -1,84 +1,55 @@
 package com.jerboa.ui.components.home
 
-import android.app.Activity
+import android.content.Context
 import android.util.Log
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Bookmarks
-import androidx.compose.material.icons.filled.Email
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.outlined.Bookmarks
-import androidx.compose.material.icons.outlined.Email
-import androidx.compose.material.icons.outlined.Home
-import androidx.compose.material.icons.outlined.Person
-import androidx.compose.material.icons.outlined.Search
+import androidx.compose.material.icons.outlined.Add
+import androidx.compose.material3.DrawerState
 import androidx.compose.material3.DrawerValue
-import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.FabPosition
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.ModalDrawerSheet
 import androidx.compose.material3.ModalNavigationDrawer
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.TopAppBarScrollBehavior
 import androidx.compose.material3.rememberDrawerState
-import androidx.compose.material3.surfaceColorAtElevation
+import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import androidx.navigation.compose.currentBackStackEntryAsState
-import com.google.accompanist.navigation.animation.AnimatedNavHost
-import com.google.accompanist.navigation.animation.rememberAnimatedNavController
 import com.jerboa.R
-import com.jerboa.datatypes.api.GetUnreadCountResponse
+import com.jerboa.VoteType
+import com.jerboa.closeDrawer
+import com.jerboa.db.Account
 import com.jerboa.db.AccountViewModel
-import com.jerboa.db.AppSettings
 import com.jerboa.db.AppSettingsViewModel
-import com.jerboa.nav.bottomNavDefaultRoute
-import com.jerboa.nav.bottomNavIsHome
-import com.jerboa.nav.bottomNavIsInbox
-import com.jerboa.nav.bottomNavIsProfile
-import com.jerboa.nav.bottomNavIsSaved
-import com.jerboa.nav.bottomNavIsSearch
-import com.jerboa.nav.bottomNavSelectHome
-import com.jerboa.nav.bottomNavSelectInbox
-import com.jerboa.nav.bottomNavSelectProfile
-import com.jerboa.nav.bottomNavSelectSaved
-import com.jerboa.nav.bottomNavSelectSearch
-import com.jerboa.nav.communityListScreen
-import com.jerboa.nav.feedScreen
-import com.jerboa.nav.inboxScreen
-import com.jerboa.nav.noEnterTransition
-import com.jerboa.nav.noPopExitTransition
-import com.jerboa.nav.profileScreen
-import com.jerboa.nav.showLogin
-import com.jerboa.ui.components.comment.edit.CommentEditViewModel
-import com.jerboa.ui.components.comment.reply.CommentReplyViewModel
-import com.jerboa.ui.components.common.InboxIconAndBadge
+import com.jerboa.fetchInitialData
+import com.jerboa.loginFirstToast
+import com.jerboa.scrollToTop
+import com.jerboa.ui.components.common.BottomAppBarAll
 import com.jerboa.ui.components.common.getCurrentAccount
-import com.jerboa.ui.components.community.list.CommunityListViewModel
-import com.jerboa.ui.components.inbox.InboxViewModel
-import com.jerboa.ui.components.person.PersonProfileViewModel
+import com.jerboa.ui.components.common.getPostViewMode
+import com.jerboa.ui.components.post.PostListings
 import com.jerboa.ui.components.post.edit.PostEditViewModel
-import com.jerboa.unreadCountTotal
+import kotlinx.coroutines.CoroutineScope
 
-@OptIn(ExperimentalAnimationApi::class)
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun HomeActivity(
     navController: NavController,
@@ -88,34 +59,16 @@ fun HomeActivity(
     postEditViewModel: PostEditViewModel,
     appSettingsViewModel: AppSettingsViewModel,
     showVotingArrowsInListView: Boolean,
-    communityListViewModel: CommunityListViewModel,
-    inboxViewModel: InboxViewModel,
-    commentReplyViewModel: CommentReplyViewModel,
-    personProfileViewModel: PersonProfileViewModel,
-    commentEditViewModel: CommentEditViewModel,
-    appSettings: AppSettings?,
 ) {
     Log.d("jerboa", "got to home activity")
 
     val scope = rememberCoroutineScope()
+    val postListState = rememberLazyListState()
     val snackbarHostState = remember { SnackbarHostState() }
-    val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
+    val drawerState = rememberDrawerState(DrawerValue.Closed)
+    val scrollBehavior = TopAppBarDefaults.enterAlwaysScrollBehavior(rememberTopAppBarState())
     val ctx = LocalContext.current
     val account = getCurrentAccount(accountViewModel)
-
-    var selectedBottomNavBarType by rememberSaveable { mutableStateOf(BottomNavBarType.Home) }
-    val bottomNavController = rememberAnimatedNavController()
-    val navBackStackEntry by bottomNavController.currentBackStackEntryAsState()
-    LaunchedEffect(navBackStackEntry) {
-        selectedBottomNavBarType = when {
-            navBackStackEntry?.bottomNavIsHome() ?: false -> BottomNavBarType.Home
-            navBackStackEntry?.bottomNavIsSearch() ?: false -> BottomNavBarType.Search
-            navBackStackEntry?.bottomNavIsInbox() ?: false -> BottomNavBarType.Inbox
-            navBackStackEntry?.bottomNavIsSaved() ?: false -> BottomNavBarType.Saved
-            navBackStackEntry?.bottomNavIsProfile() ?: false -> BottomNavBarType.Profile
-            else -> BottomNavBarType.Home
-        }
-    }
 
     ModalNavigationDrawer(
         drawerState = drawerState,
@@ -125,7 +78,6 @@ fun HomeActivity(
                     MainDrawer(
                         siteViewModel = siteViewModel,
                         navController = navController,
-                        bottomNavController = bottomNavController,
                         accountViewModel = accountViewModel,
                         homeViewModel = homeViewModel,
                         scope = scope,
@@ -137,95 +89,79 @@ fun HomeActivity(
         },
         content = {
             Scaffold(
+                modifier = Modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
                 snackbarHost = { SnackbarHost(snackbarHostState) },
+                topBar = {
+                    MainTopBar(
+                        scope = scope,
+                        postListState = postListState,
+                        drawerState = drawerState,
+                        homeViewModel = homeViewModel,
+                        appSettingsViewModel = appSettingsViewModel,
+                        account = account,
+                        ctx = ctx,
+                        navController = navController,
+                        scrollBehavior = scrollBehavior,
+                    )
+                },
                 content = { padding ->
-                    AnimatedNavHost(
-                        navController = bottomNavController,
-                        startDestination = bottomNavDefaultRoute(),
-                        modifier = Modifier.padding(bottom = padding.calculateBottomPadding()),
-                        enterTransition = { noEnterTransition },
-                        popExitTransition = { noPopExitTransition },
+                    MainPostListingsContent(
+                        padding = padding,
+                        homeViewModel = homeViewModel,
+                        siteViewModel = siteViewModel,
+                        postEditViewModel = postEditViewModel,
+                        appSettingsViewModel = appSettingsViewModel,
+                        account = account,
+                        ctx = ctx,
+                        navController = navController,
+                        postListState = postListState,
+                        showVotingArrowsInListView = showVotingArrowsInListView,
+                    )
+                },
+                floatingActionButtonPosition = FabPosition.End,
+                floatingActionButton = {
+                    FloatingActionButton(
+                        onClick = {
+                            account?.also {
+                                navController.navigate("createPost")
+                            } ?: run {
+                                loginFirstToast(ctx)
+                            }
+                        },
                     ) {
-                        feedScreen(
-                            navController = navController,
-                            homeViewModel = homeViewModel,
-                            accountViewModel = accountViewModel,
-                            siteViewModel = siteViewModel,
-                            postEditViewModel = postEditViewModel,
-                            appSettingsViewModel = appSettingsViewModel,
-                            drawerState = drawerState,
-                            appSettings = appSettings,
-                        )
-
-                        communityListScreen(
-                            navController = navController,
-                            accountViewModel = accountViewModel,
-                            communityListViewModel = communityListViewModel,
-                            siteViewModel = siteViewModel,
-                        )
-
-                        inboxScreen(
-                            navController = navController,
-                            inboxViewModel = inboxViewModel,
-                            accountViewModel = accountViewModel,
-                            homeViewModel = homeViewModel,
-                            commentReplyViewModel = commentReplyViewModel,
-                            siteViewModel = siteViewModel,
-                            account = account,
-                            ctx = ctx,
-                        )
-
-                        profileScreen(
-                            navController = navController,
-                            personProfileViewModel = personProfileViewModel,
-                            accountViewModel = accountViewModel,
-                            commentEditViewModel = commentEditViewModel,
-                            commentReplyViewModel = commentReplyViewModel,
-                            postEditViewModel = postEditViewModel,
-                            appSettingsViewModel = appSettingsViewModel,
-                            siteViewModel = siteViewModel,
-                            account = account,
-                            appSettings = appSettings,
+                        Icon(
+                            imageVector = Icons.Outlined.Add,
+                            contentDescription = stringResource(R.string.floating_createPost),
                         )
                     }
                 },
                 bottomBar = {
-                    BottomNavBar(
+                    BottomAppBarAll(
                         showBottomNav = appSettingsViewModel.appSettings.value?.showBottomNav,
-                        selectedType = selectedBottomNavBarType,
+                        screen = "home",
                         unreadCounts = homeViewModel.unreadCountResponse,
-                        onClickHome = {
-                            selectedBottomNavBarType = BottomNavBarType.Home
-                            bottomNavController.bottomNavSelectHome()
-                        },
-                        onClickSearch = {
-                            selectedBottomNavBarType = BottomNavBarType.Search
-                            bottomNavController.bottomNavSelectSearch()
+                        onClickProfile = {
+                            account?.id?.also {
+                                navController.navigate(route = "profile/$it")
+                            } ?: run {
+                                loginFirstToast(ctx)
+                            }
                         },
                         onClickInbox = {
                             account?.also {
-                                selectedBottomNavBarType = BottomNavBarType.Inbox
-                                bottomNavController.bottomNavSelectInbox()
+                                navController.navigate(route = "inbox")
                             } ?: run {
-                                navController.showLogin()
+                                loginFirstToast(ctx)
                             }
                         },
                         onClickSaved = {
                             account?.id?.also {
-                                selectedBottomNavBarType = BottomNavBarType.Saved
-                                bottomNavController.bottomNavSelectSaved(it)
+                                navController.navigate(route = "profile/$it?saved=${true}")
                             } ?: run {
-                                navController.showLogin()
+                                loginFirstToast(ctx)
                             }
                         },
-                        onClickProfile = {
-                            account?.id?.also {
-                                selectedBottomNavBarType = BottomNavBarType.Profile
-                                bottomNavController.bottomNavSelectProfile(it)
-                            } ?: run {
-                                navController.showLogin()
-                            }
-                        },
+                        navController = navController,
                     )
                 },
             )
@@ -233,97 +169,269 @@ fun HomeActivity(
     )
 }
 
-enum class BottomNavBarType {
-    Home, Search, Inbox, Saved, Profile,
-}
-
 @Composable
-fun BottomNavBar(
-    selectedType: BottomNavBarType,
-    unreadCounts: GetUnreadCountResponse? = null,
-    showBottomNav: Boolean? = true,
-    onClickHome: () -> Unit,
-    onClickSearch: () -> Unit,
-    onClickSaved: () -> Unit,
-    onClickProfile: () -> Unit,
-    onClickInbox: () -> Unit,
+fun MainPostListingsContent(
+    homeViewModel: HomeViewModel,
+    siteViewModel: SiteViewModel,
+    postEditViewModel: PostEditViewModel,
+    account: Account?,
+    ctx: Context,
+    navController: NavController,
+    padding: PaddingValues,
+    postListState: LazyListState,
+    appSettingsViewModel: AppSettingsViewModel,
+    showVotingArrowsInListView: Boolean,
 ) {
-    val totalUnreads = unreadCounts?.let { unreadCountTotal(it) }
-
-    if (showBottomNav == true) {
-        // Check for preview mode
-        if (LocalContext.current is Activity) {
-            val window = (LocalContext.current as Activity).window
-            val colorScheme = MaterialTheme.colorScheme
-
-            DisposableEffect(Unit) {
-                window.navigationBarColor = colorScheme.surfaceColorAtElevation(3.dp).toArgb()
-
-                onDispose {
-                    window.navigationBarColor = colorScheme.background.toArgb()
-                }
-            }
-        }
-
-        NavigationBar {
-            for (type in BottomNavBarType.values()) {
-                val selected = type == selectedType
-                NavigationBarItem(
-                    icon = {
-                        InboxIconAndBadge(
-                            iconBadgeCount = if (type == BottomNavBarType.Inbox) totalUnreads else null,
-                            icon = if (selected) {
-                                when (type) {
-                                    BottomNavBarType.Home -> Icons.Filled.Home
-                                    BottomNavBarType.Search -> Icons.Filled.Search
-                                    BottomNavBarType.Inbox -> Icons.Filled.Email
-                                    BottomNavBarType.Saved -> Icons.Filled.Bookmarks
-                                    BottomNavBarType.Profile -> Icons.Filled.Person
-                                }
-                            } else {
-                                when (type) {
-                                    BottomNavBarType.Home -> Icons.Outlined.Home
-                                    BottomNavBarType.Search -> Icons.Outlined.Search
-                                    BottomNavBarType.Inbox -> Icons.Outlined.Email
-                                    BottomNavBarType.Saved -> Icons.Outlined.Bookmarks
-                                    BottomNavBarType.Profile -> Icons.Outlined.Person
-                                }
-                            },
-                            contentDescription = stringResource(
-                                when (type) {
-                                    BottomNavBarType.Home -> R.string.bottomBar_home
-                                    BottomNavBarType.Search -> R.string.bottomBar_search
-                                    BottomNavBarType.Inbox -> R.string.bottomBar_inbox
-                                    BottomNavBarType.Saved -> R.string.bottomBar_bookmarks
-                                    BottomNavBarType.Profile -> R.string.bottomBar_profile
-                                },
-                            ),
-                        )
-                    },
-                    label = {
-                        Text(
-                            text = stringResource(
-                                when (type) {
-                                    BottomNavBarType.Home -> R.string.bottomBar_label_home
-                                    BottomNavBarType.Search -> R.string.bottomBar_label_search
-                                    BottomNavBarType.Inbox -> R.string.bottomBar_label_inbox
-                                    BottomNavBarType.Saved -> R.string.bottomBar_label_bookmarks
-                                    BottomNavBarType.Profile -> R.string.bottomBar_label_profile
-                                },
-                            ),
-                            color = MaterialTheme.colorScheme.onSurface,
-                        )
-                    },
-                    selected = selected,
-                    onClick = when (type) {
-                        BottomNavBarType.Home -> onClickHome
-                        BottomNavBarType.Search -> onClickSearch
-                        BottomNavBarType.Inbox -> onClickInbox
-                        BottomNavBarType.Saved -> onClickSaved
-                        BottomNavBarType.Profile -> onClickProfile
-                    },
+    PostListings(
+        listState = postListState,
+        padding = padding,
+        posts = homeViewModel.posts,
+        taglines = siteViewModel.siteRes?.taglines,
+        postViewMode = getPostViewMode(appSettingsViewModel),
+        onUpvoteClick = { postView ->
+            homeViewModel.likePost(
+                voteType = VoteType.Upvote,
+                postView = postView,
+                account = account,
+                ctx = ctx,
+            )
+        },
+        onDownvoteClick = { postView ->
+            homeViewModel.likePost(
+                voteType = VoteType.Downvote,
+                postView = postView,
+                account = account,
+                ctx = ctx,
+            )
+        },
+        onPostClick = { postView ->
+            navController.navigate(route = "post/${postView.post.id}")
+        },
+        onSaveClick = { postView ->
+            account?.also { acct ->
+                homeViewModel.savePost(
+                    postView = postView,
+                    account = acct,
+                    ctx = ctx,
                 )
             }
+        },
+        onBlockCommunityClick = {
+            account?.also { acct ->
+                homeViewModel.blockCommunity(
+                    community = it,
+                    account = acct,
+                    ctx = ctx,
+                )
+            }
+        },
+        onBlockCreatorClick = {
+            account?.also { acct ->
+                homeViewModel.blockCreator(
+                    creator = it,
+                    account = acct,
+                    ctx = ctx,
+                )
+            }
+        },
+        onEditPostClick = { postView ->
+            postEditViewModel.initialize(postView)
+            navController.navigate("postEdit")
+        },
+        onDeletePostClick = { postView ->
+            account?.also { acct ->
+                homeViewModel.deletePost(
+                    postView = postView,
+                    account = acct,
+                    ctx = ctx,
+                )
+            }
+        },
+        onReportClick = { postView ->
+            navController.navigate("postReport/${postView.post.id}")
+        },
+        onCommunityClick = { community ->
+            navController.navigate(route = "community/${community.id}")
+        },
+        onPersonClick = { personId ->
+            navController.navigate(route = "profile/$personId")
+        },
+        onSwipeRefresh = {
+            homeViewModel.fetchPosts(
+                account = account,
+                clear = true,
+                ctx = ctx,
+            )
+        },
+        loading = homeViewModel.loading.value &&
+            homeViewModel.page.value == 1 &&
+            homeViewModel.posts.isNotEmpty(),
+        isScrolledToEnd = {
+            if (homeViewModel.posts.size > 0) {
+                homeViewModel.fetchPosts(
+                    account = account,
+                    nextPage = true,
+                    ctx = ctx,
+                )
+            }
+        },
+        account = account,
+        showVotingArrowsInListView = showVotingArrowsInListView,
+        enableDownVotes = siteViewModel.siteRes?.site_view?.local_site?.enable_downvotes ?: true,
+        showAvatar = siteViewModel.siteRes?.my_user?.local_user_view?.local_user?.show_avatars ?: true,
+    )
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun MainDrawer(
+    siteViewModel: SiteViewModel,
+    navController: NavController,
+    accountViewModel: AccountViewModel,
+    homeViewModel: HomeViewModel,
+    scope: CoroutineScope,
+    ctx: Context,
+    drawerState: DrawerState,
+) {
+    val accounts = accountViewModel.allAccounts.value
+    val account = getCurrentAccount(accountViewModel)
+
+    Drawer(
+        myUserInfo = siteViewModel.siteRes?.my_user,
+        unreadCounts = homeViewModel.unreadCountResponse,
+        accountViewModel = accountViewModel,
+        navController = navController,
+        isOpen = drawerState.isOpen,
+        onSwitchAccountClick = { acct ->
+            accountViewModel.removeCurrent()
+            accountViewModel.setCurrent(acct.id)
+
+            fetchInitialData(
+                account = acct,
+                siteViewModel = siteViewModel,
+                homeViewModel = homeViewModel,
+            )
+
+            closeDrawer(scope, drawerState)
+        },
+        onSignOutClick = {
+            accounts?.also { accts ->
+                account?.also {
+                    accountViewModel.delete(it)
+                    val updatedList = accts.toMutableList()
+                    updatedList.remove(it)
+
+                    if (updatedList.isNotEmpty()) {
+                        accountViewModel.setCurrent(updatedList[0].id)
+                    }
+                    fetchInitialData(
+                        account = updatedList.getOrNull(0),
+                        siteViewModel = siteViewModel,
+                        homeViewModel = homeViewModel,
+                    )
+
+                    closeDrawer(scope, drawerState)
+                }
+            }
+        },
+        onClickListingType = { listingType ->
+            homeViewModel.fetchPosts(
+                account = account,
+                clear = true,
+                changeListingType = listingType,
+                ctx = ctx,
+            )
+            closeDrawer(scope, drawerState)
+        },
+        onCommunityClick = { community ->
+            navController.navigate(route = "community/${community.id}")
+            closeDrawer(scope, drawerState)
+        },
+        onClickProfile = {
+            account?.id?.also {
+                navController.navigate(route = "profile/$it")
+                closeDrawer(scope, drawerState)
+            }
+        },
+        onClickSaved = {
+            account?.id?.also {
+                navController.navigate(route = "profile/$it?saved=${true}")
+                closeDrawer(scope, drawerState)
+            }
+        },
+        onClickInbox = {
+            account?.also {
+                navController.navigate(route = "inbox")
+            } ?: run {
+                loginFirstToast(ctx)
+            }
+            closeDrawer(scope, drawerState)
+        },
+        onClickSettings = {
+            navController.navigate(route = "settings")
+            closeDrawer(scope, drawerState)
+        },
+        onClickCommunities = {
+            navController.navigate(route = "communityList")
+            closeDrawer(scope, drawerState)
+        },
+    )
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun MainTopBar(
+    scope: CoroutineScope,
+    postListState: LazyListState,
+    drawerState: DrawerState,
+    homeViewModel: HomeViewModel,
+    appSettingsViewModel: AppSettingsViewModel,
+    account: Account?,
+    ctx: Context,
+    navController: NavController,
+    scrollBehavior: TopAppBarScrollBehavior,
+) {
+    Column {
+        HomeHeader(
+            scope = scope,
+            scrollBehavior = scrollBehavior,
+            drawerState = drawerState,
+            navController = navController,
+            selectedSortType = homeViewModel.sortType.value,
+            selectedListingType = homeViewModel.listingType.value,
+            selectedPostViewMode = getPostViewMode(appSettingsViewModel),
+            onClickSortType = { sortType ->
+                scrollToTop(scope, postListState)
+                homeViewModel.fetchPosts(
+                    account = account,
+                    clear = true,
+                    changeSortType = sortType,
+                    ctx = ctx,
+                )
+            },
+            onClickListingType = { listingType ->
+                scrollToTop(scope, postListState)
+                homeViewModel.fetchPosts(
+                    account = account,
+                    clear = true,
+                    changeListingType = listingType,
+                    ctx = ctx,
+                )
+            },
+            onClickPostViewMode = {
+                appSettingsViewModel.updatedPostViewMode(it.ordinal)
+            },
+            onClickRefresh = {
+                scrollToTop(scope, postListState)
+                homeViewModel.fetchPosts(
+                    account = account,
+                    clear = true,
+                    ctx = ctx,
+                )
+            },
+        )
+        if (homeViewModel.loading.value) {
+            LinearProgressIndicator(modifier = Modifier.fillMaxWidth())
         }
     }
 }
