+++ app/src/main/java/com/jerboa/ui/components/inbox/InboxActivity.kt
@@ -23,10 +23,12 @@ import androidx.navigation.NavController
 import com.jerboa.*
 import com.jerboa.db.Account
 import com.jerboa.db.AccountViewModel
+import com.jerboa.db.AppSettingsViewModel
 import com.jerboa.ui.components.comment.mentionnode.CommentMentionNode
 import com.jerboa.ui.components.comment.reply.CommentReplyViewModel
 import com.jerboa.ui.components.comment.reply.ReplyItem
 import com.jerboa.ui.components.comment.replynode.CommentReplyNode
+import com.jerboa.ui.components.common.BottomAppBarAll
 import com.jerboa.ui.components.common.getCurrentAccount
 import com.jerboa.ui.components.common.simpleVerticalScrollbar
 import com.jerboa.ui.components.home.HomeViewModel
@@ -39,6 +41,7 @@ import kotlinx.coroutines.launch
 @Composable
 fun InboxActivity(
     navController: NavController,
+    appSettingsViewModel: AppSettingsViewModel,
     inboxViewModel: InboxViewModel,
     homeViewModel: HomeViewModel,
     accountViewModel: AccountViewModel,
@@ -107,6 +110,33 @@ fun InboxActivity(
                 account = account,
                 scope = scope,
                 siteViewModel = siteViewModel,
+            )
+        },
+        bottomBar = {
+            BottomAppBarAll(
+                showBottomNav = appSettingsViewModel.appSettings.value?.showBottomNav,
+                screen = "inbox",
+                unreadCounts = homeViewModel.unreadCountResponse,
+                onClickProfile = {
+                    account?.id?.also {
+                        navController.navigate(route = "profile/$it")
+                    }
+                },
+                onClickInbox = {
+                    account?.also {
+                        navController.navigate(route = "inbox")
+                    } ?: run {
+                        loginFirstToast(ctx)
+                    }
+                },
+                onClickSaved = {
+                    account?.id?.also {
+                        navController.navigate(route = "profile/$it?saved=${true}")
+                    } ?: run {
+                        loginFirstToast(ctx)
+                    }
+                },
+                navController = navController,
             )
         },
     )
