+++ app/src/main/java/com/jerboa/ui/components/person/PersonProfile.kt
@@ -38,7 +38,6 @@ import com.jerboa.datatypes.SortType
 import com.jerboa.datatypes.samplePersonView
 import com.jerboa.getLocalizedSortingTypeName
 import com.jerboa.personNameShown
-import com.jerboa.ui.components.common.DefaultBackButton
 import com.jerboa.ui.components.common.DotSpacer
 import com.jerboa.ui.components.common.IconAndTextDrawerItem
 import com.jerboa.ui.components.common.ImageViewerDialog
@@ -216,7 +215,14 @@ fun PersonProfileHeader(
                 selectedSortType = selectedSortType,
             )
         },
-        navigationIcon = { DefaultBackButton(navController) },
+        navigationIcon = {
+            IconButton(onClick = { navController.popBackStack() }) {
+                Icon(
+                    Icons.Outlined.ArrowBack,
+                    contentDescription = stringResource(R.string.person_profile_back),
+                )
+            }
+        },
         actions = {
             IconButton(onClick = {
                 showSortOptions = !showSortOptions
