+++ app/src/main/java/com/jerboa/ui/components/post/create/CreatePost.kt
@@ -18,6 +18,7 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Add
 import androidx.compose.material.icons.outlined.ArrowDropDown
+import androidx.compose.material.icons.outlined.Close
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -41,7 +42,6 @@ import com.jerboa.datatypes.CommunitySafe
 import com.jerboa.datatypes.sampleCommunitySafe
 import com.jerboa.db.Account
 import com.jerboa.ui.components.common.CircularIcon
-import com.jerboa.ui.components.common.DefaultBackButton
 import com.jerboa.ui.components.common.MarkdownTextField
 import com.jerboa.ui.components.common.PickImage
 import com.jerboa.ui.theme.ICON_SIZE
@@ -83,8 +83,17 @@ fun CreatePostHeader(
             }
         },
         navigationIcon = {
-            // Todo add are you sure cancel dialog
-            DefaultBackButton(navController)
+            IconButton(
+                onClick = {
+                    navController.popBackStack()
+                },
+            ) {
+                // Todo add are you sure cancel dialog
+                Icon(
+                    Icons.Outlined.Close,
+                    contentDescription = stringResource(R.string.create_post_close),
+                )
+            }
         },
     )
 }
