+++ app/src/main/java/com/jerboa/ui/components/post/create/CreatePostActivity.kt
@@ -73,7 +73,6 @@ fun CreatePostActivity(
             }
         }
     }
-
     Surface(color = MaterialTheme.colorScheme.background) {
         Scaffold(
             topBar = {
